web: fix client workspaces cicular dependency issue
