+++ client/observability-client/src/types/@opentelemetry/api/index.d.ts
