+++ client/observability-client/src/types/@opentelemetry/core/index.d.ts
