+++ client/observability-client/src/types/@opentelemetry/exporter-trace-otlp-http/index.d.ts
