+++ client/observability-client/src/types/@opentelemetry/resources/index.d.ts
