+++ client/observability-client/src/types/@opentelemetry/sdk-trace-base/index.d.ts
