+++ client/observability-client/src/types/@opentelemetry/semantic-conventions/index.d.ts
