+++ client/observability-client/tsconfig.json
@@ -7,7 +7,7 @@
     "rootDir": ".",
     "outDir": "out",
     "paths": {
-      "*": ["../shared/src/types/*", "*"],
+      "*": ["./src/types/*", "*"],
     },
   },
   "include": ["**/*", ".*"],
