+++ client/search-ui/src/components/CodeExcerpt.tsx
@@ -10,10 +10,8 @@ import { catchError, filter } from 'rxjs/operators'
 import { HoverMerged } from '@sourcegraph/client-api'
 import { DOMFunctions, findPositionsFromEvents, Hoverifier } from '@sourcegraph/codeintellify'
 import { asError, ErrorLike, isDefined, isErrorLike, highlightNodeMultiline } from '@sourcegraph/common'
-import { HighlightLineRange } from '@sourcegraph/search'
 import { ActionItemAction } from '@sourcegraph/shared/src/actions/ActionItem'
 import { ViewerId } from '@sourcegraph/shared/src/api/viewerTypes'
-import { HighlightResponseFormat } from '@sourcegraph/shared/src/graphql-operations'
 import { HoverContext } from '@sourcegraph/shared/src/hover/HoverOverlay.types'
 import { Repo } from '@sourcegraph/shared/src/util/url'
 import { Icon, Code } from '@sourcegraph/wildcard'
@@ -25,15 +23,6 @@ export interface Shape {
     left?: number
     bottom?: number
     right?: number
-}
-
-export interface FetchFileParameters {
-    repoName: string
-    commitID: string
-    filePath: string
-    disableTimeout?: boolean
-    ranges: HighlightLineRange[]
-    format?: HighlightResponseFormat
 }
 
 interface Props extends Repo {
