+++ client/vscode/src/webview/search-panel/SearchResultsView.tsx
@@ -10,15 +10,9 @@ import {
     getUserSearchContextNamespaces,
     QueryState,
 } from '@sourcegraph/search'
-import {
-    IEditor,
-    SearchBox,
-    StreamingProgress,
-    StreamingSearchResultsList,
-    FetchFileParameters,
-} from '@sourcegraph/search-ui'
+import { IEditor, SearchBox, StreamingProgress, StreamingSearchResultsList } from '@sourcegraph/search-ui'
 import { wrapRemoteObservable } from '@sourcegraph/shared/src/api/client/api/common'
-import { fetchHighlightedFileLineRanges } from '@sourcegraph/shared/src/backend/file'
+import { FetchFileParameters, fetchHighlightedFileLineRanges } from '@sourcegraph/shared/src/backend/file'
 import { collectMetrics } from '@sourcegraph/shared/src/search/query/metrics'
 import {
     appendContextFilter,
