+++ client/web/src/notebooks/blocks/query/NotebookQueryBlock.tsx
@@ -10,12 +10,12 @@ import { Hoverifier } from '@sourcegraph/codeintellify'
 import { SearchContextProps } from '@sourcegraph/search'
 import {
     StreamingSearchResultsList,
-    FetchFileParameters,
     CodeMirrorQueryInput,
     changeListener,
     createDefaultSuggestions,
 } from '@sourcegraph/search-ui'
 import { ActionItemAction } from '@sourcegraph/shared/src/actions/ActionItem'
+import { FetchFileParameters } from '@sourcegraph/shared/src/backend/file'
 import { editorHeight } from '@sourcegraph/shared/src/components/CodeMirrorEditor'
 import { ExtensionsControllerProps } from '@sourcegraph/shared/src/extensions/controller'
 import { HoverContext } from '@sourcegraph/shared/src/hover/HoverOverlay.types'
