+++ client/web/src/search/results/StreamingSearchResults.tsx
@@ -7,7 +7,8 @@ import { Observable } from 'rxjs'
 
 import { asError } from '@sourcegraph/common'
 import { QueryUpdate, SearchContextProps, SearchMode } from '@sourcegraph/search'
-import { FetchFileParameters, StreamingProgress, StreamingSearchResultsList } from '@sourcegraph/search-ui'
+import { StreamingProgress, StreamingSearchResultsList } from '@sourcegraph/search-ui'
+import { FetchFileParameters } from '@sourcegraph/shared/src/backend/file'
 import { FilePrefetcher } from '@sourcegraph/shared/src/components/PrefetchableFile'
 import { ExtensionsControllerProps } from '@sourcegraph/shared/src/extensions/controller'
 import { SearchPatternType } from '@sourcegraph/shared/src/graphql-operations'
